data_collectin_service/requirements.txt
celery
redis
requests
sqlalchemy
requests
requests-mock
python-dotenv
psycopg2
pytest